backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==4.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.9.2
python-dotenv==0.19.0